% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sinaplot.R
\name{sinaplot}
\alias{sinaplot}
\title{sinaplot}
\usage{
sinaplot(x, groups, method = "density", scale = TRUE, yFraction = 0.02,
  neighbLimit = 1, adjust = 3/4, xSpread = 0.1, labels = NULL,
  plot = TRUE, ...)
}
\arguments{
\item{x}{numeric vector of values to be plotted}

\item{groups}{vector of length x}

\item{method}{choose the method to spread the samples within the same
neighbourhood along the x-axis. Available methods: "density",
"neighbourhood". See \code{Details}.}

\item{scale}{logical. When set to \code{TRUE} x-coordinate widths across all
groups are scaled based on the densiest are in the plot. Default: \code{TRUE}}

\item{yFraction}{binning factor. The range of the values in \code{x} are
binned in windows of length \code{(max(x) - min(x)) * yFraction}. Samples
within the same bin belong to the same "neighbourhood".}

\item{neighbLimit}{if the samples within the same y-axis bin are more than
neighbLimit, the samples's X coordinates will be adjusted}

\item{adjust}{adjusts the bandwidth of the density kernel when
\code{method = "density"} (see \code{\link[stats]{density}}).}

\item{xSpread}{tuning parameter that adjusts the spread of the samples within
the same neighbourhood along the x-axis when \code{method = "neighbourhood"}.
Accepts values between 0 and 1.}

\item{labels}{optional arguments that controls the labels along the x-axis.
Must be of length \code{unique(groups)}.}

\item{plot}{logical. When \code{TRUE} the sinaplot is produced, otherwise the
function returns the new sample coordinates. Default: \code{TRUE}}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link{par}}).}
}
\value{
if \code{plot = FALSE} a list is returned containing a data.frame
with the new sample coordinates and a vector of length x that corresponds to
each sample's group.
}
\description{
blah blah blah blah blah blah blah blah blah blah blah blah
}
\details{
kati
\itemize{
 \item{\bold{"details"}: blah blah}
 \item{"neighbourhood": blsd }
}
}
\examples{
x <- c(rnorm(200, 4, 1), rnorm(200, 5, 2), rnorm(200, 6, 1.5))
groups <- c(rep("Cond1", 200), rep("Cond2", 200), rep("Cond3", 200))

sinaplot(x, groups)
sinaplot(x, groups, scale = FALSE)
sinaplot(x, groups, scale = FALSE, adjust = 1/6)
sinaplot(x, groups, scale = FALSE, adjust = 3)
}

