% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sinaplot.R
\name{sinaplot}
\alias{sinaplot}
\title{sinaplot}
\usage{
sinaplot(x, groups, labels = NULL, plot = TRUE, relScaling = T,
  ypercentage = 0.02, xpercentage = 0.1, neighbLimit = 1, adjust = 1,
  method = "neighbours", ...)
}
\arguments{
\item{x}{numeric vector of values to be plotted}

\item{groups}{vector of length x}

\item{labels}{optional}

\item{plot}{logical. Default: TRUE}

\item{neighbLimit}{blah blah}

\item{adjust}{blah blah blah}

\item{method}{blah blah blah blah}

\item{...}{}
}
\value{
blah blah blah blah
}
\description{
blah blah blah blah blah blah blah blah blah blah blah blah
}

